feat: moved checkbox to separate component
